rename: 디렉토리 구조 및 이름 변경
