+++ src/main/java/umc/meme/auth/global/exception/handler/ControllerAdvice.java
@@ -1,10 +1,10 @@
-package umc.meme.auth.global.exception;
+package umc.meme.auth.global.exception.handler;
 
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import umc.meme.auth.global.common.BaseResponseDto;
 import umc.meme.auth.global.common.status.ErrorStatus;
-import umc.meme.auth.global.exception.handler.AuthException;
+import umc.meme.auth.global.exception.AuthException;
 
 @RestControllerAdvice
 public class ControllerAdvice {
